corrigindo bugs

+++ app/assets/javascripts/show.js
@@ -2,7 +2,7 @@ function initMap() {
   var lat1 = document.getElementById('lat').value;
   var lon1 = document.getElementById('lon').value;
   
-  var myLatLng = {lat: parseInt(lat1), lng: parseInt(lon1)};
+  var myLatLng = {lat: parseFloat(lat1), lng: parseFloat(lon1)};
 
   var map = new google.maps.Map(document.getElementById('map'), {
     zoom: 16,
